Week1_PerimeterAssignmentRunner_Janna
This is my version of the homework assignment from week 1. This version doesn't have the bugs that I noticed in the other GitHub versions of this homework whereby the name of the largest file is wrong, or the FileResource() dialog box only appears 2 or 3 times and doesn't allow for multiple selections. I tested this code on 2 versions of the Coursera Quiz and passed both with 100%.